requirements.txt
decorator==4.3.0
glove-python==0.1.0
networkx==2.2
numpy==1.15.4
Optunity==1.1.1
Pillow==5.3.0
psycopg2==2.7.6.1
pycodestyle==2.4.0
PyYAML==3.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
torch==0.4.1
torchvision==0.2.1
